Split up more code.
