+++ libs/render-wgpu/buffer/terrain_geo/Cargo.toml
@@ -20,3 +20,4 @@ frame_graph = { path = "../../../render-wgpu/common/frame_graph" }
 geodesy = { path = "../../../geodesy" }
 geometry = { path = "../../../render-shared/common/geometry" }
 gpu = { path = "../../../render-wgpu/common/gpu" }
+universe = { path = "../../../universe" }
